error handling

+++ streamlit_app.py
@@ -59,34 +59,40 @@ def generate_audio(video_description):
 
 
 def store_audio_file(mp3_bytes_io, video_file_name):
-    user_id = get_user_id()
-    print(f"store_audio_file() for user_id: {user_id}, video: {video_file_name}")
-    mp3_file_name = os.path.splitext(video_file_name)[0]+'.mp3'
-    tmp_mp3_file = f"/tmp/{mp3_file_name}"
+    try:
+        user_id = get_user_id()
+        print(f"store_audio_file() for user_id: {user_id}, video: {video_file_name}")
+        mp3_file_name = os.path.splitext(video_file_name)[0]+'.mp3'
+        tmp_mp3_file = f"/tmp/{mp3_file_name}"
 
-    cloud_mp3_file = f"{CLOUD_STORAGE_BUCKET}/audio/{user_id}/{mp3_file_name}"
-    print(f"Writing {cloud_mp3_file} to cloud storage")
+        cloud_mp3_file = f"{CLOUD_STORAGE_BUCKET}/audio/{user_id}/{mp3_file_name}"
+        print(f"Writing {cloud_mp3_file} to cloud storage")
 
-    with open(tmp_mp3_file, "wb") as tmp_file:
-        tmp_file.write(mp3_bytes_io.getbuffer())
+        with open(tmp_mp3_file, "wb") as tmp_file:
+            tmp_file.write(mp3_bytes_io.getbuffer())
 
-    file_storage_conn = st.connection('gcs', type=FilesConnection)
-    with file_storage_conn.open(cloud_mp3_file, "wb") as gcs_file:
-        gcs_file.write(mp3_bytes_io.getbuffer())
-
-    os.remove(tmp_mp3_file)
+        file_storage_conn = st.connection('gcs', type=FilesConnection)
+        with file_storage_conn.open(cloud_mp3_file, "wb") as gcs_file:
+            gcs_file.write(mp3_bytes_io.getbuffer())
+    except Exception as e:
+        print(f"Exception: {e}")
+    finally:
+        if os.path.exists(tmp_mp3_file):
+            os.remove(tmp_mp3_file)
     return cloud_mp3_file
 
 
 def store_video_details(video_details):
-    print("begin store_video_details")
-    key_dict = json.loads(st.secrets.FIREBASE_KEY)
+    try:
+        print("begin store_video_details")
+        key_dict = json.loads(st.secrets.FIREBASE_KEY)
 
-    credentials = service_account.Credentials.from_service_account_info(key_dict)
-    firestore_db = firestore.Client(credentials=credentials, project="eyehear-firebase")
-    collection = firestore_db.collection("videos")
-    video_document = collection.add(document_data=video_details)
-
+        credentials = service_account.Credentials.from_service_account_info(key_dict)
+        firestore_db = firestore.Client(credentials=credentials, project="eyehear-firebase")
+        collection = firestore_db.collection("videos")
+        video_document = collection.add(document_data=video_details)
+    except Exception as e:
+        print(f"Exception: {e}")
     print("completed store_video_details")
 
 model = create_gemini_model()
@@ -95,7 +101,8 @@ st.write(
     "Proof of concept"
 )
 st.write(
-    "For now we'll manually upload an mp4 doorbell video.  Ultimately, this would listen for incoming doorbell events."
+    """For now we'll manually upload an mp4 doorbell video.  Ultimately, this would 
+    listen for incoming doorbell events, and automatically read transcriptions aloud."""
 )
 
 uploaded_file = st.file_uploader("Upload doorbell video", type=['mp4'])
@@ -123,6 +130,7 @@ if uploaded_file is not None:
     store_video_details(response_data)
 
 if st.button("Example video", type="primary"):
+    # just send the video to Gemini and generate audio
     example_url = "https://github.com/ThatOrJohn/eye-hear-streamlit/raw/main/examples/Ring_FrontDoor_202408061842.mp4"
     response = model.generate_content(example_url)
     response_data = json.loads(response.text)
